Initial commit for converting app.schema
