+++ src/middlewared/middlewared/pytest/unit/plugins/apps/test_normalize_and_validate.py
@@ -1,6 +1,6 @@
 import pytest
 
-from middlewared.plugins.apps.schema_normalization import AppSchemaService
+from middlewared.plugins.apps.schema_normalization_old import AppSchemaService
 from middlewared.pytest.unit.middleware import Middleware
